Cleanup model and add interface.
